+++ pytato/__init__.py
@@ -29,7 +29,6 @@ from pytato.array import (
 
         make_dict_of_named_arrays,
         make_placeholder, make_size_param, make_data_wrapper,
-        make_distributed_recv, make_distributed_send,
         AdditionalOutput,
 
         matmul, roll, transpose, stack, reshape, concatenate,
@@ -46,6 +45,8 @@ from pytato.array import (
         logical_or, logical_and, logical_not,
 
         )
+
+from pytato.distributed import make_distributed_send, make_distributed_recv
 
 from pytato.target.loopy.codegen import generate_loopy
 from pytato.target import Target
